Removed default constants in BasicCellPainter
The constants that had been returned when the
actual value for a cell was "null" made it
impossible to simply skip certain rendering steps
based on this value (e.g. one could not skip
drawing a shape by returning a "null" Stroke).
Now, the defaults are only set in the constructor,
and may be changed to return "null" values in
order to skip the corresponding rendering steps.